+++ src/actions/plip.js
@@ -260,3 +260,8 @@ export const fetchingPlipRelatedInfo = isFetching => ({
 export const clearPlipInfo = () => ({
   type: "PLIP_CLEAR_INFO",
 });
+
+export const plipsSingInfoFetched = ({ signInfo }) => ({
+  type: "PLIP_PLIPS_SIGN_INFO_FETCHED",
+  payload: { signInfo },
+});