+++ src/sagas/plip.js
@@ -2,6 +2,12 @@ import { takeEvery, takeLatest } from "redux-saga";
 import { cancelled, call, spawn, put, select } from "redux-saga/effects";
 
 import {
+  isNil,
+  pluck,
+  zip,
+} from "ramda";
+
+import {
   homeSceneKey,
   isDev,
   isUnauthorized,
@@ -44,6 +50,7 @@ import {
   plipsNationwideFetchError,
   plipsStatewideFetchError,
   plipsCitywideFetchError,
+  plipsSingInfoFetched,
   plipJustSigned,
   nationwidePlipsRefreshError,
   statewidePlipsRefreshError,
@@ -78,6 +85,8 @@ import { profileScreenForCurrentUser } from "./navigation";
 import { validateLocalWallet } from "./wallet";
 
 import LibCrypto from "mudamos-libcrypto";
+
+const isNotNil = x => !isNil(x);
 
 const buildSignMessage = ({ user, plip }) => [
   user.name,
@@ -175,10 +184,7 @@ export function* fetchNationWidePlipsSaga({ mudamosWebApi }) {
       yield put(fetchingNationwidePlips(true));
       const response = yield call(fetchPlips, { mudamosWebApi });
 
-      if (!response.nextPage && response.plips && response.plips.length) {
-        response.plips = [...response.plips, { key: "LINKS" }];
-      }
-
+      response.plips = yield call(processPlipsResponse, response);
       yield put(nationwidePlipsFetched(response));
 
       yield put(fetchingNationwidePlips(false));
@@ -201,10 +207,7 @@ function* fetchNationwidePlipsNextPageSaga({ mudamosWebApi }) {
       yield put(fetchingNextNationwidePlipsPage(true));
       const response = yield call(fetchPlips, { page, mudamosWebApi });
 
-      if (!response.nextPage) {
-        response.plips = [...(response.plips || []), { key: "LINKS" }];
-      }
-
+      response.plips = yield call(processPlipsResponseForNextPage, response);
       yield put(plipsAppendNationwidePlips(response));
     } catch (e) {
       logError(e);
@@ -222,10 +225,7 @@ function* refreshNationwidePlipsSaga({ mudamosWebApi }) {
       yield put(isRefreshingNationwidePlips(true));
       const response = yield call(fetchPlips, { page: 1, mudamosWebApi });
 
-      if (!response.nextPage && response.plips && response.plips.length) {
-        response.plips = [...response.plips, { key: "LINKS" }];
-      }
-
+      response.plips = yield call(processPlipsResponse, response);
       yield put(nationwidePlipsFetched(response));
     } catch (e) {
       logError(e);
@@ -250,10 +250,7 @@ function* fetchStatewidePlipsSaga({ mudamosWebApi }) {
         uf: filters.state.uf,
       });
 
-      if (!response.nextPage && response.plips && response.plips.length) {
-        response.plips = [...response.plips, { key: "LINKS" }];
-      }
-
+      response.plips = yield call(processPlipsResponse, response);
       yield put(statewidePlipsFetched(response));
     } catch (e) {
       logError(e);
@@ -277,10 +274,7 @@ function* fetchStatewidePlipsNextPageSaga({ mudamosWebApi }) {
         uf: filters.state.uf,
       });
 
-      if (!response.nextPage) {
-        response.plips = [...(response.plips || []), { key: "LINKS" }];
-      }
-
+      response.plips = yield call(processPlipsResponseForNextPage, response);
       yield put(plipsAppendStatewidePlips(response));
     } catch (e) {
       logError(e);
@@ -304,10 +298,7 @@ function* refreshStatewidePlipsSaga({ mudamosWebApi }) {
         uf: filters.state.uf,
       });
 
-      if (!response.nextPage && response.plips && response.plips.length) {
-        response.plips = [...response.plips, { key: "LINKS" }];
-      }
-
+      response.plips = yield call(processPlipsResponse, response);
       yield put(statewidePlipsFetched(response));
     } catch (e) {
       logError(e);
@@ -330,10 +321,7 @@ function* fetchCitywidePlipsSaga({ mudamosWebApi }) {
         cityId: filters.city.id,
       });
 
-      if (!response.nextPage && response.plips && response.plips.length) {
-        response.plips = [...response.plips, { key: "LINKS" }];
-      }
-
+      response.plips = yield call(processPlipsResponse, response);
       yield put(citywidePlipsFetched(response));
     } catch (e) {
       logError(e);
@@ -357,10 +345,7 @@ function* fetchCitywidePlipsNextPageSaga({ mudamosWebApi }) {
         cityId: filters.city.id,
       });
 
-      if (!response.nextPage) {
-        response.plips = [...(response.plips || []), { key: "LINKS" }];
-      }
-
+      response.plips = yield call(processPlipsResponseForNextPage, response);
       yield put(plipsAppendCitywidePlips(response));
     } catch (e) {
       logError(e);
@@ -384,10 +369,7 @@ function* refreshCitywidePlipsSaga({ mudamosWebApi }) {
         cityId: filters.city.id,
       });
 
-      if (!response.nextPage && response.plips && response.plips.length) {
-        response.plips = [...response.plips, { key: "LINKS" }];
-      }
-
+      response.plips = yield call(processPlipsResponse, response);
       yield put(citywidePlipsFetched(response));
     } catch (e) {
       logError(e);
@@ -415,6 +397,30 @@ function* changePlipsFilterCity() {
 
 function* fetchPlips({ mudamosWebApi, page = 1, uf, cityId }) {
   return yield call(mudamosWebApi.listPlips, { page, uf, cityId });
+}
+
+/*
+ * After fetching or refreshing we must add the plips links accordingly.
+ */
+function processPlipsResponse({ plips, nextPage }) {
+  // Add the mudamos links to the end only if there are plips
+  if (!nextPage && plips && plips.length) {
+    return [...plips, { key: "LINKS" }];
+  }
+
+  return plips;
+}
+
+/*
+ * After fetching the next page we must add the plips links accordingly.
+ */
+function processPlipsResponseForNextPage({ plips, nextPage }) {
+  // Add the mudamos links to the end
+  if (!nextPage) {
+    return [...(plips || []), { key: "LINKS" }];
+  }
+
+  return plips;
 }
 
 function* fetchPlipRelatedInfo({ mobileApi }) {
@@ -614,6 +620,35 @@ function* invalidateWalletAndNavigate(params = {}) {
   yield put(profileStateMachine(params));
 }
 
+function* fetchPlipsSignInfoSaga({ mobileApi }) {
+  const actions = [
+    "PLIPS_NATIONWIDE_FETCHED",
+    "PLIPS_STATEWIDE_FETCHED",
+    "PLIPS_CITYWIDE_FETCHED",
+    "PLIPS_APPEND_NATIONWIDE_PLIPS",
+    "PLIPS_APPEND_STATEWIDE_PLIPS",
+    "PLIPS_APPEND_CITYWIDE_PLIPS",
+  ];
+
+  yield takeEvery(actions, function* ({ payload }) {
+    try {
+      const { plips } = payload;
+      const plipIds = pluck("id", plips || []).filter(isNotNil);
+      if (!plipIds.length) return;
+
+      const results = yield plipIds.map(id => call(mobileApi.plipSignInfo, id))
+
+      const signInfo = zip(plipIds, results).reduce((memo, [id, result]) => {
+        memo[id] = result.info;
+        return memo;
+      }, {});
+      yield put(plipsSingInfoFetched({ signInfo }));
+    } catch(e) {
+      logError(e, { tag: "fetchPlipsSignInfoSaga" });
+    }
+  });
+}
+
 export default function* plipSaga({ mobileApi, mudamosWebApi, walletStore, apiError }) {
   yield spawn(fetchFilteredPlips);
   yield spawn(fetchFilteredPlipsNextPage);
@@ -633,4 +668,5 @@ export default function* plipSaga({ mobileApi, mudamosWebApi, walletStore, apiEr
   yield spawn(updatePlipSignInfoSaga, { mobileApi });
   yield spawn(fetchPlipSignersSaga, { mobileApi });
   yield spawn(fetchPlipRelatedInfo, { mobileApi });
+  yield spawn(fetchPlipsSignInfoSaga, { mobileApi });
 }